+++ packages/site/examples/case/graphDemos/demo/largeGraph.js
@@ -722,7 +722,7 @@ const processNodesEdges = (
   // let maxCount = 0;
   edges.forEach((edge) => {
     // to avoid the dulplicated id to nodes
-    if (!edge.id) edge.id = `edge-${uniqueId()}`;
+    if (!edge.id) edge.id = uniqueId('edge');
     else if (edge.id.split('-')[0] !== 'edge') edge.id = `edge-${edge.id}`;
     // TODO: delete the following line after the queried data is correct
     if (!currentNodeMap[edge.source] || !currentNodeMap[edge.target]) {
@@ -785,9 +785,8 @@ const processNodesEdges = (
     const arrowWidth = Math.max(size / 2 + 2, 3);
     const arrowLength = 10;
     const arrowBeging = targetNode.size + arrowLength;
-    let arrowPath = `M ${arrowBeging},0 L ${arrowBeging + arrowLength},-${arrowWidth} L ${
-      arrowBeging + arrowLength
-    },${arrowWidth} Z`;
+    let arrowPath = `M ${arrowBeging},0 L ${arrowBeging + arrowLength},-${arrowWidth} L ${arrowBeging + arrowLength
+      },${arrowWidth} Z`;
     let d = targetNode.size / 2 + arrowLength;
     if (edge.source === edge.target) {
       edge.type = 'loop';
@@ -811,11 +810,11 @@ const processNodesEdges = (
       lineDash,
       endArrow: arrowPath
         ? {
-            path: arrowPath,
-            d,
-            fill: stroke,
-            strokeOpacity: 0,
-          }
+          path: arrowPath,
+          d,
+          fill: stroke,
+          strokeOpacity: 0,
+        }
         : false,
     };
     edge.labelCfg = {
@@ -1084,7 +1083,7 @@ const getMixedGraph = (
       const vedge = {
         source: edge.source,
         target: targetClusterId,
-        id: `edge-${uniqueId()}`,
+        id: uniqueId('edge'),
         label: '',
       };
       edges.push(vedge);
@@ -1093,7 +1092,7 @@ const getMixedGraph = (
       const vedge = {
         target: edge.target,
         source: sourceClusterId,
-        id: `edge-${uniqueId()}`,
+        id: uniqueId('edge'),
         label: '',
       };
       edges.push(vedge);
@@ -1152,7 +1151,7 @@ const generateNeighbors = (centerNodeModel, step, maxNeighborNumPerNode = 5) => 
   const neighborNum = Math.ceil(Math.random() * maxNeighborNumPerNode);
   for (let i = 0; i < neighborNum; i++) {
     const neighborNode = {
-      id: uniqueId(),
+      id: uniqueId('node'),
       clusterId,
       level: 0,
       colorSet: centerNodeModel.colorSet,
@@ -1162,7 +1161,7 @@ const generateNeighbors = (centerNodeModel, step, maxNeighborNumPerNode = 5) => 
     const source = dire ? centerId : neighborNode.id;
     const target = dire ? neighborNode.id : centerId;
     const neighborEdge = {
-      id: uniqueId(),
+      id: uniqueId('edge'),
       source,
       target,
       label: `${source}-${target}`,
@@ -1198,7 +1197,7 @@ const getExtractNodeMixedGraph = (
       if (!aggregatedNodeMap[targetClusterId].expanded) {
         // did not expand, create an virtual edge fromt he extract node to the cluster
         currentUnproccessedData.edges.push({
-          id: uniqueId(),
+          id: uniqueId('edge'),
           source: extractNodeId,
           target: targetClusterId,
         });
@@ -1211,7 +1210,7 @@ const getExtractNodeMixedGraph = (
       if (!aggregatedNodeMap[sourceClusterId].expanded) {
         // did not expand, create an virtual edge fromt he extract node to the cluster
         currentUnproccessedData.edges.push({
-          id: uniqueId(),
+          id: uniqueId('edge'),
           target: extractNodeId,
           source: sourceClusterId,
         });
@@ -1473,7 +1472,7 @@ fetch('https://gw.alipayobjects.com/os/antvdemo/assets/data/relations.json')
         ...clusterEdge,
         size: Math.log(clusterEdge.count),
         label: '',
-        id: `edge-${uniqueId()}`,
+        id: uniqueId('edge'),
       };
       if (cedge.source === cedge.target) {
         cedge.type = 'loop';
@@ -1486,7 +1485,7 @@ fetch('https://gw.alipayobjects.com/os/antvdemo/assets/data/relations.json')
 
     data.edges.forEach((edge) => {
       edge.label = `${edge.source}-${edge.target}`;
-      edge.id = `edge-${uniqueId()}`;
+      edge.id = uniqueId('edge');
     });
 
     currentUnproccessedData = aggregatedData;