+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - feat: Annotation supports icon events callbacks;
 - feat: Annotation supports defaultBegin position configuration for new annotation cards;
 - perf: Annotation updated automatically when graph data changed and graph item visiblity changed;
+- fix: Destroy legend canvas when the plugin is destroyed, closes: #3931;
 
 ### 4.7.2